refactor(tests): extend HostedInspection to remaining view tests (#145)
Cover the 4 view test files with unwrapped .inspect() calls that
weren't enumerated in issue #145's file list but match its
"~23 files not touched by PR #127" scope, completing the refactor.

- ProviderListViewTests (19 sites)
- ProviderDetailViewTests (26 sites) — split into
  ProviderDetailViewTests + ProviderDetailViewFieldTests to respect
  SwiftLint type_body_length after HostedInspection wrapping added
  indentation. Form field and pre-fill tests moved to the new file;
  structural, mode, button, and section tests stay in the original.
- MedicalRecordDetailViewActionTests (4 sites) — converted inline
  ViewHosting.host/expel pairs to HostedInspection.inspect
- ExportWarningDialogTests (10 sites)

Tests: 1641/1644 pass, 3 skipped, 0 fail. Coverage: 80.92%
(up from 80.77%).
