+++ ios/FamilyMedicalApp/FamilyMedicalAppTests/Views/Provider/ProviderDetailViewTests.swift
@@ -3,6 +3,9 @@ import Testing
 import ViewInspector
 @testable import FamilyMedicalApp
 
+/// Structural, mode, button, and section tests for ProviderDetailView.
+///
+/// Field presence and pre-fill tests live in ProviderDetailViewFieldTests.
 @MainActor
 struct ProviderDetailViewTests {
     // MARK: - Test Data
@@ -36,8 +39,10 @@ struct ProviderDetailViewTests {
         let person = try makeTestPerson()
         let view = ProviderDetailView(person: person) { _ in true }
 
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.NavigationStack.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.NavigationStack.self)
+        }
     }
 
     @Test
@@ -45,8 +50,10 @@ struct ProviderDetailViewTests {
         let person = try makeTestPerson()
         let view = ProviderDetailView(person: person) { _ in true }
 
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.Form.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.Form.self)
+        }
     }
 
     @Test
@@ -54,8 +61,10 @@ struct ProviderDetailViewTests {
         let person = try makeTestPerson()
         let view = ProviderDetailView(person: person) { _ in true }
 
-        let inspected = try view.inspect()
-        _ = try inspected.find(text: "Is this a person or a practice? Fill in their name, organization, or both.")
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(text: "Is this a person or a practice? Fill in their name, organization, or both.")
+        }
     }
 
     @Test
@@ -64,11 +73,13 @@ struct ProviderDetailViewTests {
         let existingProvider = makeProvider()
         let view = ProviderDetailView(person: person, existingProvider: existingProvider) { _ in true }
 
-        let inspected = try view.inspect()
-        // In edit mode, the hint section should not be present
-        let hintText = try? inspected
-            .find(text: "Is this a person or a practice? Fill in their name, organization, or both.")
-        #expect(hintText == nil)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            // In edit mode, the hint section should not be present
+            let hintText = try? inspected
+                .find(text: "Is this a person or a practice? Fill in their name, organization, or both.")
+            #expect(hintText == nil)
+        }
     }
 
     // MARK: - Edit Mode Tests
@@ -79,8 +90,10 @@ struct ProviderDetailViewTests {
         let existingProvider = makeProvider()
         let view = ProviderDetailView(person: person, existingProvider: existingProvider) { _ in true }
 
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.NavigationStack.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.NavigationStack.self)
+        }
     }
 
     @Test
@@ -89,163 +102,10 @@ struct ProviderDetailViewTests {
         let existingProvider = makeProvider()
         let view = ProviderDetailView(person: person, existingProvider: existingProvider) { _ in true }
 
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.Form.self)
-    }
-
-    // MARK: - Form Field Tests
-
-    @Test
-    func formHasNameField() throws {
-        let person = try makeTestPerson()
-        let view = ProviderDetailView(person: person) { _ in true }
-
-        let inspected = try view.inspect()
-        let form = try inspected.find(ViewType.Form.self)
-        // Provider Information section has Name, Organization, Specialty
-        _ = try form.find(ViewType.TextField.self) {
-            try $0.labelView().text().string() == "Name"
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.Form.self)
         }
-    }
-
-    @Test
-    func formHasOrganizationField() throws {
-        let person = try makeTestPerson()
-        let view = ProviderDetailView(person: person) { _ in true }
-
-        let inspected = try view.inspect()
-        let form = try inspected.find(ViewType.Form.self)
-        _ = try form.find(ViewType.TextField.self) {
-            try $0.labelView().text().string() == "Organization"
-        }
-    }
-
-    @Test
-    func formHasSpecialtyField() throws {
-        let person = try makeTestPerson()
-        let view = ProviderDetailView(person: person) { _ in true }
-
-        let inspected = try view.inspect()
-        let form = try inspected.find(ViewType.Form.self)
-        _ = try form.find(ViewType.TextField.self) {
-            try $0.labelView().text().string() == "Specialty"
-        }
-    }
-
-    @Test
-    func formHasPhoneField() throws {
-        let person = try makeTestPerson()
-        let view = ProviderDetailView(person: person) { _ in true }
-
-        let inspected = try view.inspect()
-        let form = try inspected.find(ViewType.Form.self)
-        _ = try form.find(ViewType.TextField.self) {
-            try $0.labelView().text().string() == "Phone"
-        }
-    }
-
-    @Test
-    func formHasAddressField() throws {
-        let person = try makeTestPerson()
-        let view = ProviderDetailView(person: person) { _ in true }
-
-        let inspected = try view.inspect()
-        let form = try inspected.find(ViewType.Form.self)
-        _ = try form.find(ViewType.TextField.self) {
-            try $0.labelView().text().string() == "Address"
-        }
-    }
-
-    @Test
-    func formHasNotesField() throws {
-        let person = try makeTestPerson()
-        let view = ProviderDetailView(person: person) { _ in true }
-
-        let inspected = try view.inspect()
-        let form = try inspected.find(ViewType.Form.self)
-        _ = try form.find(ViewType.TextEditor.self)
-    }
-
-    // MARK: - Pre-filled Data Tests
-
-    @Test
-    func formPreFillsNameFromExistingProvider() throws {
-        let person = try makeTestPerson()
-        let existingProvider = makeProvider(name: "Dr. Jane Doe")
-        let view = ProviderDetailView(person: person, existingProvider: existingProvider) { _ in true }
-
-        let inspected = try view.inspect()
-        let nameField = try inspected.find(ViewType.TextField.self) {
-            try $0.labelView().text().string() == "Name"
-        }
-        let inputValue = try nameField.input()
-        #expect(inputValue == "Dr. Jane Doe")
-    }
-
-    @Test
-    func formPreFillsOrganizationFromExistingProvider() throws {
-        let person = try makeTestPerson()
-        let existingProvider = makeProvider(name: "Dr. Smith", organization: "City Hospital")
-        let view = ProviderDetailView(person: person, existingProvider: existingProvider) { _ in true }
-
-        let inspected = try view.inspect()
-        let orgField = try inspected.find(ViewType.TextField.self) {
-            try $0.labelView().text().string() == "Organization"
-        }
-        let inputValue = try orgField.input()
-        #expect(inputValue == "City Hospital")
-    }
-
-    @Test
-    func formPreFillsSpecialtyFromExistingProvider() throws {
-        let person = try makeTestPerson()
-        let existingProvider = makeProvider(specialty: "Neurology")
-        let view = ProviderDetailView(person: person, existingProvider: existingProvider) { _ in true }
-
-        let inspected = try view.inspect()
-        let specialtyField = try inspected.find(ViewType.TextField.self) {
-            try $0.labelView().text().string() == "Specialty"
-        }
-        let inputValue = try specialtyField.input()
-        #expect(inputValue == "Neurology")
-    }
-
-    @Test
-    func formPreFillsPhoneFromExistingProvider() throws {
-        let person = try makeTestPerson()
-        let existingProvider = makeProvider(phone: "555-1234")
-        let view = ProviderDetailView(person: person, existingProvider: existingProvider) { _ in true }
-
-        let inspected = try view.inspect()
-        let phoneField = try inspected.find(ViewType.TextField.self) {
-            try $0.labelView().text().string() == "Phone"
-        }
-        let inputValue = try phoneField.input()
-        #expect(inputValue == "555-1234")
-    }
-
-    @Test
-    func formPreFillsAddressFromExistingProvider() throws {
-        let person = try makeTestPerson()
-        let existingProvider = makeProvider(address: "456 Oak Ave")
-        let view = ProviderDetailView(person: person, existingProvider: existingProvider) { _ in true }
-
-        let inspected = try view.inspect()
-        let addressField = try inspected.find(ViewType.TextField.self) {
-            try $0.labelView().text().string() == "Address"
-        }
-        let inputValue = try addressField.input()
-        #expect(inputValue == "456 Oak Ave")
-    }
-
-    @Test
-    func formPreFillsNotesFromExistingProvider() throws {
-        let person = try makeTestPerson()
-        let existingProvider = makeProvider(notes: "Very helpful")
-        let view = ProviderDetailView(person: person, existingProvider: existingProvider) { _ in true }
-
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.TextEditor.self)
     }
 
     // MARK: - Button Tests
@@ -255,8 +115,10 @@ struct ProviderDetailViewTests {
         let person = try makeTestPerson()
         let view = ProviderDetailView(person: person) { _ in true }
 
-        let inspected = try view.inspect()
-        _ = try inspected.find(button: "Save")
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(button: "Save")
+        }
     }
 
     @Test
@@ -264,8 +126,10 @@ struct ProviderDetailViewTests {
         let person = try makeTestPerson()
         let view = ProviderDetailView(person: person) { _ in true }
 
-        let inspected = try view.inspect()
-        _ = try inspected.find(button: "Cancel")
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(button: "Cancel")
+        }
     }
 
     // MARK: - Section Tests
@@ -275,11 +139,13 @@ struct ProviderDetailViewTests {
         let person = try makeTestPerson()
         let view = ProviderDetailView(person: person) { _ in true }
 
-        let inspected = try view.inspect()
-        let form = try inspected.find(ViewType.Form.self)
-        // Create mode: section 0 is the hint, section 1 is Provider Information
-        _ = try form.section(0)
-        _ = try form.section(1)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            let form = try inspected.find(ViewType.Form.self)
+            // Create mode: section 0 is the hint, section 1 is Provider Information
+            _ = try form.section(0)
+            _ = try form.section(1)
+        }
     }
 
     @Test
@@ -287,10 +153,12 @@ struct ProviderDetailViewTests {
         let person = try makeTestPerson()
         let view = ProviderDetailView(person: person) { _ in true }
 
-        let inspected = try view.inspect()
-        let form = try inspected.find(ViewType.Form.self)
-        // Create mode: section 0=hint, 1=provider info, 2=contact
-        _ = try form.section(2)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            let form = try inspected.find(ViewType.Form.self)
+            // Create mode: section 0=hint, 1=provider info, 2=contact
+            _ = try form.section(2)
+        }
     }
 
     @Test
@@ -298,10 +166,12 @@ struct ProviderDetailViewTests {
         let person = try makeTestPerson()
         let view = ProviderDetailView(person: person) { _ in true }
 
-        let inspected = try view.inspect()
-        let form = try inspected.find(ViewType.Form.self)
-        // Create mode: section 0=hint, 1=provider info, 2=contact, 3=notes
-        _ = try form.section(3)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            let form = try inspected.find(ViewType.Form.self)
+            // Create mode: section 0=hint, 1=provider info, 2=contact, 3=notes
+            _ = try form.section(3)
+        }
     }
 
     @Test
@@ -310,13 +180,15 @@ struct ProviderDetailViewTests {
         let existingProvider = makeProvider()
         let view = ProviderDetailView(person: person, existingProvider: existingProvider) { _ in true }
 
-        let inspected = try view.inspect()
-        let form = try inspected.find(ViewType.Form.self)
-        // Edit mode: conditional hint section is absent (Optional at index 0),
-        // so real sections start at index 1 in ViewInspector
-        _ = try form.section(1)
-        _ = try form.section(2)
-        _ = try form.section(3)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            let form = try inspected.find(ViewType.Form.self)
+            // Edit mode: conditional hint section is absent (Optional at index 0),
+            // so real sections start at index 1 in ViewInspector
+            _ = try form.section(1)
+            _ = try form.section(2)
+            _ = try form.section(3)
+        }
     }
 
     // MARK: - Empty Field Tests
@@ -326,12 +198,14 @@ struct ProviderDetailViewTests {
         let person = try makeTestPerson()
         let view = ProviderDetailView(person: person) { _ in true }
 
-        let inspected = try view.inspect()
-        let nameField = try inspected.find(ViewType.TextField.self) {
-            try $0.labelView().text().string() == "Name"
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            let nameField = try inspected.find(ViewType.TextField.self) {
+                try $0.labelView().text().string() == "Name"
+            }
+            let inputValue = try nameField.input()
+            #expect(inputValue.isEmpty)
         }
-        let inputValue = try nameField.input()
-        #expect(inputValue.isEmpty)
     }
 
     @Test
@@ -347,10 +221,12 @@ struct ProviderDetailViewTests {
         )
         let view = ProviderDetailView(person: person, existingProvider: existingProvider) { _ in true }
 
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.Form.self)
-        // Verify all text fields are present
-        let textFields = inspected.findAll(ViewType.TextField.self)
-        #expect(textFields.count == 5)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.Form.self)
+            // Verify all text fields are present
+            let textFields = inspected.findAll(ViewType.TextField.self)
+            #expect(textFields.count == 5)
+        }
     }
 }
