+++ ios/FamilyMedicalApp/FamilyMedicalAppTests/Views/Provider/ProviderListViewTests.swift
@@ -36,8 +36,10 @@ struct ProviderListViewTests {
         let viewModel = makeViewModel(person: person)
         let view = ProviderListView(person: person, viewModel: viewModel)
 
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.Group.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.Group.self)
+        }
     }
 
     @Test
@@ -47,8 +49,10 @@ struct ProviderListViewTests {
         // No providers, not loading -> ContentUnavailableView
         let view = ProviderListView(person: person, viewModel: viewModel)
 
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.ContentUnavailableView.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.ContentUnavailableView.self)
+        }
     }
 
     @Test
@@ -58,8 +62,10 @@ struct ProviderListViewTests {
         viewModel.isLoading = true
         let view = ProviderListView(person: person, viewModel: viewModel)
 
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.ProgressView.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.ProgressView.self)
+        }
     }
 
     @Test
@@ -75,8 +81,10 @@ struct ProviderListViewTests {
         await viewModel.loadProviders()
 
         let view = ProviderListView(person: person, viewModel: viewModel)
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.List.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.List.self)
+        }
     }
 
     @Test
@@ -90,11 +98,13 @@ struct ProviderListViewTests {
         await viewModel.loadProviders()
 
         let view = ProviderListView(person: person, viewModel: viewModel)
-        let inspected = try view.inspect()
-        let list = try inspected.find(ViewType.List.self)
-        let forEach = try list.forEach(0)
-        // Verify at least one row exists
-        _ = try forEach.button(0)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            let list = try inspected.find(ViewType.List.self)
+            let forEach = try list.forEach(0)
+            // Verify at least one row exists
+            _ = try forEach.button(0)
+        }
     }
 
     @Test
@@ -104,9 +114,11 @@ struct ProviderListViewTests {
         let view = ProviderListView(person: person, viewModel: viewModel)
 
         // Verify the view renders with searchable modifier
-        let inspected = try view.inspect()
-        // The searchable modifier is applied to the Group, verify rendering works
-        _ = try inspected.find(ViewType.Group.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            // The searchable modifier is applied to the Group, verify rendering works
+            _ = try inspected.find(ViewType.Group.self)
+        }
     }
 
     @Test
@@ -115,10 +127,12 @@ struct ProviderListViewTests {
         let viewModel = makeViewModel(person: person)
         let view = ProviderListView(person: person, viewModel: viewModel)
 
-        let inspected = try view.inspect()
-        // Find the toolbar button with the plus image
-        _ = try inspected.find(ViewType.Image.self) { image in
-            try image.actualImage().name() == "plus"
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            // Find the toolbar button with the plus image
+            _ = try inspected.find(ViewType.Image.self) { image in
+                try image.actualImage().name() == "plus"
+            }
         }
     }
 
@@ -129,9 +143,11 @@ struct ProviderListViewTests {
         viewModel.errorMessage = "Unable to load providers."
 
         let view = ProviderListView(person: person, viewModel: viewModel)
-        let inspected = try view.inspect()
-        // View renders successfully even in error state
-        _ = try inspected.find(ViewType.Group.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            // View renders successfully even in error state
+            _ = try inspected.find(ViewType.Group.self)
+        }
         #expect(viewModel.errorMessage == "Unable to load providers.")
     }
 
@@ -145,8 +161,10 @@ struct ProviderListViewTests {
         await viewModel.loadProviders()
 
         let view = ProviderListView(person: person, viewModel: viewModel)
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.Group.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.Group.self)
+        }
         #expect(viewModel.errorMessage != nil)
     }
 
@@ -161,9 +179,11 @@ struct ProviderListViewTests {
         await viewModel.loadProviders()
 
         let view = ProviderListView(person: person, viewModel: viewModel)
-        let inspected = try view.inspect()
-        // Find text matching the provider displayString
-        _ = try inspected.find(text: "Dr. Adams at Health Clinic")
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            // Find text matching the provider displayString
+            _ = try inspected.find(text: "Dr. Adams at Health Clinic")
+        }
     }
 
     @Test
@@ -177,8 +197,10 @@ struct ProviderListViewTests {
         await viewModel.loadProviders()
 
         let view = ProviderListView(person: person, viewModel: viewModel)
-        let inspected = try view.inspect()
-        _ = try inspected.find(text: "Dermatology")
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(text: "Dermatology")
+        }
     }
 
     @Test
@@ -188,8 +210,10 @@ struct ProviderListViewTests {
         let view = ProviderListView(person: person, viewModel: viewModel)
 
         // View renders with confirmation dialog modifier
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.Group.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.Group.self)
+        }
     }
 
     @Test
@@ -197,8 +221,10 @@ struct ProviderListViewTests {
         let person = try makeTestPerson()
         // Initialize without providing viewModel to test default initialization
         let view = ProviderListView(person: person)
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.Group.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.Group.self)
+        }
     }
 
     @Test
@@ -219,13 +245,15 @@ struct ProviderListViewTests {
         await viewModel.loadProviders()
 
         let view = ProviderListView(person: person, viewModel: viewModel)
-        let inspected = try view.inspect()
-        let list = try inspected.find(ViewType.List.self)
-        let forEach = try list.forEach(0)
-        // Verify all 3 rows rendered
-        _ = try forEach.button(0)
-        _ = try forEach.button(1)
-        _ = try forEach.button(2)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            let list = try inspected.find(ViewType.List.self)
+            let forEach = try list.forEach(0)
+            // Verify all 3 rows rendered
+            _ = try forEach.button(0)
+            _ = try forEach.button(1)
+            _ = try forEach.button(2)
+        }
     }
 
     @Test
@@ -239,8 +267,10 @@ struct ProviderListViewTests {
         await viewModel.loadProviders()
 
         let view = ProviderListView(person: person, viewModel: viewModel)
-        let inspected = try view.inspect()
-        _ = try inspected.find(text: "Sunrise Medical Center")
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(text: "Sunrise Medical Center")
+        }
     }
 
     @Test
@@ -253,9 +283,11 @@ struct ProviderListViewTests {
         await viewModel.loadProviders()
 
         let view = ProviderListView(person: person, viewModel: viewModel)
-        let inspected = try view.inspect()
-        // When providers exist, should show List not ContentUnavailableView
-        _ = try inspected.find(ViewType.List.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            // When providers exist, should show List not ContentUnavailableView
+            _ = try inspected.find(ViewType.List.self)
+        }
     }
 
     @Test
@@ -264,8 +296,10 @@ struct ProviderListViewTests {
         let viewModel = makeViewModel(person: person)
         let view = ProviderListView(person: person, viewModel: viewModel)
 
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.Group.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.Group.self)
+        }
     }
 
     @Test
@@ -274,9 +308,11 @@ struct ProviderListViewTests {
         let viewModel = makeViewModel(person: person)
         let view = ProviderListView(person: person, viewModel: viewModel)
 
-        let inspected = try view.inspect()
-        let unavailable = try inspected.find(ViewType.ContentUnavailableView.self)
-        _ = unavailable
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            let unavailable = try inspected.find(ViewType.ContentUnavailableView.self)
+            _ = unavailable
+        }
     }
 
     @Test
@@ -288,7 +324,9 @@ struct ProviderListViewTests {
         // (since filteredProviders is empty) and ProgressView overlay appear
         let view = ProviderListView(person: person, viewModel: viewModel)
 
-        let inspected = try view.inspect()
-        _ = try inspected.find(ViewType.ProgressView.self)
+        try HostedInspection.inspect(view) { view in
+            let inspected = try view.inspect()
+            _ = try inspected.find(ViewType.ProgressView.self)
+        }
     }
 }
